deepfake_detection/requirements.txt
django
torch
torchvision
efficientnet-pytorch
timm
opencv-python
mediapipe
numpy
Pillow
scikit-learn
matplotlib
seaborn
python-dotenv
django-crispy-forms
crispy-bootstrap5
whitenoise
mlflow
